+++ src/StatePulse.NET/Configuration/LifetimeEnum.cs
@@ -1,6 +1,0 @@
-﻿namespace StatePulse.Net.Configuration;
-public enum LifetimeEnum
-{
-    Singleton,
-    Scoped
-}