+++ src/StatePulse.NET/Engine/IPulseGlobalTracker.cs
@@ -1,0 +1,8 @@
+﻿namespace StatePulse.Net.Engine;
+public interface IPulseGlobalTracker
+{
+    public int ActivePulsars { get; }
+    public void Register(IStatePulse pulsar);
+    public void UnRegister(IStatePulse pulsar);
+    public event EventHandler? onAfterCleanUp;
+}