+++ src/StatePulse.NET/IEffectValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace StatePulse.Net;
+public interface IEffectValidator<in TAction, TEffect>
+    where TAction : IAction
+    where TEffect : IEffect<TAction>
+{
+    Task<bool> Validate(TAction action);
+}