+++ src/StatePulse.NET/IPulseGlobalTracker.cs
@@ -1,8 +1,0 @@
-﻿namespace StatePulse.Net;
-public interface IPulseGlobalTracker
-{
-    public int ActivePulsars { get; }
-    public void Register(IStatePulse pulsar);
-    public void UnRegister(IStatePulse pulsar);
-    public event EventHandler? onAfterCleanUp;
-}