+++ tests/StatPulse.NET.Tests/TestCases/Pulsars/Profile/Actions/ProfileCardDefineResultAction.cs
@@ -3,5 +3,5 @@
 namespace StatePulse.NET.Tests.TestCases.Pulsars.Profile.Actions;
 public record ProfileCardDefineResultAction(string Name, string Picture, string Id) : IAction
 {
-
+    public string? UnitTestStringer { get; init; }
 }